+++ introspection-engine/connectors/sql-introspection-connector/src/lib.rs
@@ -121,3 +121,20 @@ impl IntrospectionConnector for SqlIntrospectionConnector {
         Ok(introspection_result)
     }
 }
+
+trait Dedup<T: PartialEq + Clone> {
+    fn clear_duplicates(&mut self);
+}
+
+impl<T: PartialEq + Clone> Dedup<T> for Vec<T> {
+    fn clear_duplicates(&mut self) {
+        let mut already_seen = vec![];
+        self.retain(|item| match already_seen.contains(item) {
+            true => false,
+            _ => {
+                already_seen.push(item.clone());
+                true
+            }
+        })
+    }
+}
